Inquiry for Block Device Charactertics expected to succeed always
In the iscsi-test-cu.c, inquiry is being issued with block device characteristics page and if the command fails with any error other than invalid opcode, it causes a test failure. However, block device characteristics page is an optional command and the target may not be supporting this inquiry. So as a fix moved the test out of the main test code and made it part of the inquiry test suites.